+++ examples/fallenShowcase/main.js
@@ -15,6 +15,7 @@ ruitk.addElements(FallenStaging.getElements());
 
 ruitk.minimalDependacyTest();
 ruitk.dependencyHandlingTest();
+ruitk.xssTest();
 
 ruitk.append("body", ruitk.makeElements(`
     <backgroundImg>{}